// Frontend/Módulo - Consultas Amsys Navega/configuracion-bancos-navega/src/app/models/BankModel.ts
export interface BankModel {
    id?: number;
    code?: string;
    nameCompany?: string;
    bankName?: string;
    nameTransaction?: string;
    bankAccountName?: string;
    bankAccountNum?: string;
    currencyCode?: string;
    identifyingLetter?: string;
    bankAccountEbs?: string;
    bankCodeAmsys?: string;
    company?: string;
    status?: number | string;
    createdDate?: string;
}